smarthome-hub/requirements.txt
bidict==0.21.2
click==8.0.1
dnspython==1.16.0
dominate==2.6.0
eventlet==0.31.0
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-MQTT==1.1.1
Flask-SocketIO==5.1.0
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
paho-mqtt==1.5.1
python-engineio==4.2.0
python-socketio==5.3.0
six==1.16.0
visitor==0.1.3
Werkzeug==2.0.1
